Get municipalities correspondances

the correspondences indicate for each municipality existing at the start_period time, which is/are the corresponding municipality(ies) at the end_period time.

Usage

bfs_muni_correspondances(
  start_period = Sys.Date(),
  end_period = Sys.Date(),
  include_unmodified = NULL,
  include_territory_exchange = NULL,
  escape_chars = NULL,
  name_repair = "unique"
)

Arguments

start_period

Start of requested period

end_period

End of requested period

include_unmodified

Include municipalities that have not undergone any changes

include_territory_exchange

Include/exclude records that only concern territory changes.

escape_chars

Defines forbidden characters in the response and their respective escape characters. For example, with the value "/( ).(_)", all slashes in the response will be replaced with a space while periods are replaced with underscore.

name_repair

read_csv is used internally and ensures column names are "unique" using the "name_repeir" argument.

Value

a data.frame/tibble

Examples

# \donttest{
bfs_muni_correspondances(start_period = "2024-01-01", end_period = "2024-08-01")
#> # A tibble: 4 × 12
#>   InitialHistoricalCode InitialCode InitialName InitialParentHistoricalCode
#>                   <dbl>       <dbl> <chr>                             <dbl>
#> 1                 12547          64 Nürensdorf                        10081
#> 2                 12547          64 Nürensdorf                        10081
#> 3                 12673          62 Kloten                            10081
#> 4                 12673          62 Kloten                            10081
#> # ℹ 8 more variables: InitialParentName <chr>, InitialStep <dbl>,
#> #   TerminalHistoricalCode <dbl>, TerminalCode <dbl>, TerminalName <chr>,
#> #   TerminalParentHistoricalCode <dbl>, TerminalParentName <chr>,
#> #   TerminalStep <dbl>
# }